% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivity.R
\name{plotActivity}
\alias{plotActivity}
\title{Plot enzymatic activity}
\usage{
plotActivity(
  plate.tbl,
  graph.type = "Raw",
  input.plateID = c(),
  input.media = "BHET12.5",
  input.gene = c(),
  input.plasmid = "152"
)
}
\arguments{
\item{graph.type, }{restricted to four different strings ('Raw' for
raw, normalized colony.value data, 'Total' for total activity, 'Net' for net
activity, 'Net/Total' for net/total enzymatic activity)}

\item{input.plateID, }{a list of plateIDs. If multiply plate IDs are given,
the graphs will show a trendline corresponding to the mean of the
replicates across plates while showing individual replicates as data points.}

\item{input.media, }{a string which corresponds to a media condition within
the data}

\item{input.gene, }{a list of strings, can also be empty. Regardless, even if
empty, the script will automatically include "EV_ctrl", "IsPETase_ctrl", and
"FAST-PETase_ctrl".}

\item{input.plasmid}{a string which corresponds to a specific plasmids within
the data}
}
\value{
a 'ggplot2' object.
}
\description{
Generates a plot for enzymatic activity (Raw, normalized activity; Total
Activity; Net Activity; Net/Total Activity) based on the user input.
}
\details{
The data needs to be aggregated first before graphing.
}
